requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
albumentations=0.4.5=py_0
autopep8=1.5=pypi_0
bzip2=1.0.8=h0b31af3_2
ca-certificates=2020.4.5.1=hecc5488_0
cairo=1.16.0=hec6a9b0_1003
certifi=2020.4.5.1=py38h32f6830_0
cffi=1.14.0=py38h356ff06_0
chardet=3.0.4=py38h32f6830_1006
click=7.1.1=pyh8c360ce_0
cloudpickle=1.4.1=py_0
colorama=0.4.3=py_0
cryptography=2.8=py38hb761dab_2
cycler=0.10.0=py_2
cytoolz=0.10.1=py38h0b31af3_0
dask-core=2.17.2=py_0
decorator=4.4.2=py_0
ffmpeg=4.2.3=hd0c0d6a_0
fontconfig=2.13.1=h6b1039f_1001
freetype=2.10.1=h8da9a1a_0
geos=3.8.1=h4a8c4bd_0
gettext=0.19.8.1=h46ab8bc_1002
giflib=5.2.1=h0b31af3_2
glib=2.64.3=h577aef8_0
gmp=6.2.0=h4a8c4bd_2
gnutls=3.6.5=h53004b3_1002
graphite2=1.3.13=h12caacf_1001
harfbuzz=2.4.0=hd8d2a14_3
hdf5=1.10.6=nompi_h3e39495_100
icu=64.2=h6de7cb9_1
idna=2.9=py_1
imagecodecs-lite=2019.12.3=py38h64e0658_0
imageio=2.8.0=py_0
imgaug=0.4.0=py_1
jasper=1.900.1=h636a363_1006
joblib=0.14.1=py_0
jpeg=9c=h1de35cc_1001
kiwisolver=1.1.0=py38ha0d09dd_1
lame=3.100=h1de35cc_1001
libblas=3.8.0=16_openblas
libcblas=3.8.0=16_openblas
libcxx=9.0.1=1
libffi=3.2.1=h4a8c4bd_1007
libgfortran=4.0.0=2
libiconv=1.15=h0b31af3_1006
liblapack=3.8.0=16_openblas
liblapacke=3.8.0=16_openblas
libopenblas=0.3.9=h3d69b6c_0
libopencv=4.2.0=py38_6
libpng=1.6.37=hbbe82c9_1
libtiff=4.1.0=ha78913b_3
libwebp=1.0.2=hd3bf737_5
libxml2=2.9.10=h53d96d6_0
llvm-openmp=9.0.1=h28b9765_2
lz4-c=1.8.3=h6de7cb9_1001
marshmallow=3.5.0=py_0
matplotlib=3.2.1=0
matplotlib-base=3.2.1=py38h1300a51_0
mkl=2020.0=166
ncurses=6.1=h0a44026_1002
neo4j=4.0.0=pypi_0
nettle=3.4.1=h3efe00b_1002
networkx=2.4=py_1
ninja=1.10.0=ha1b3eb9_0
numpy=1.18.1=py38h1f821a2_1
olefile=0.46=py_0
opencv=4.2.0=py38_6
openh264=2.1.1=hd174df1_0
openssl=1.1.1g=h0b31af3_0
pandas=1.0.3=py38h5fc983b_0
pcre=8.44=h4a8c4bd_0
pillow=7.0.0=py38h83dc5e5_1
pip=20.0.2=py_2
pixman=0.38.0=h01d97ff_1003
platformio=4.2.1=py_0
py-opencv=4.2.0=py38h23f93f0_6
pycodestyle=2.5.0=pypi_0
pycparser=2.20=py_0
pyelftools=0.26=py_1
pyopenssl=19.1.0=py_1
pyparsing=2.4.6=py_0
pyserial=3.4=py_2
pysocks=1.7.1=py38h32f6830_1
python=3.8.2=hd5f0129_5_cpython
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.4.0=py3.8_0
pytz=2019.3=py_0
pywavelets=1.1.1=py38h65ad66c_1
pyyaml=5.3.1=py38h64e0658_0
readline=8.0=hcfe32e1_0
requests=2.23.0=pyh8c360ce_2
scikit-image=0.17.2=py38h5fc983b_0
scikit-learn=0.22.2.post1=py38h3dc85bc_0
scipy=1.4.1=py38h82752d6_0
semantic_version=2.8.2=py_0
setuptools=46.1.3=py38h32f6830_0
shapely=1.7.0=py38h5b1fbde_3
six=1.14.0=py_1
sqlite=3.30.1=h93121df_0
tabulate=0.8.7=pyh9f0ad1d_0
tifffile=2020.5.11=py_0
tk=8.6.10=hbbe82c9_0
toolz=0.10.0=py_0
torchvision=0.5.0=py38_cpu
tornado=6.0.4=py38h64e0658_1
urllib3=1.25.7=py38h32f6830_1
wheel=0.34.2=py_1
x264=1!152.20180806=h1de35cc_0
xz=5.2.4=h0b31af3_1002
yaml=0.2.4=h0b31af3_0
zlib=1.2.11=h0b31af3_1006
zstd=1.4.4=hed8d7c8_2
